backend/requirements.txt
Django>=5.2.9
psycopg2-binary
python-dotenv
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-spectacular
